Classify Unity assertion logs as errors
Recognize both native and scripting assertion mode flags before the legacy
severity mapping so assertion failures no longer appear as ordinary logs.
Extract the mode classifier to keep the flag mapping independently testable.
