+++ Assets/Tests/Editor/ConsoleLogModeClassifierTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49aa7e09ecf0044cfac0b10eb23b776f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
