+++ Packages/src/Editor/FirstPartyTools/Common/Console/ConsoleLogModeClassifier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35899c5f12ddf41ae888b52f2ab0c544
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
